+++ packages/core/src/Pipeline.ts
@@ -49,6 +49,13 @@ export default class Pipeline<Ctx extends Context, Tool extends CoreTool<any>> e
   }
 
   /**
+   * No-op implementation.
+   */
+  execute() {
+    return Promise.resolve();
+  }
+
+  /**
    * Execute all routines in order.
    */
   run<T>(initialValue?: T): Promise<any> {